+++ X4_Python_Pipe_Server/Classes/__init__.py
@@ -2,4 +2,5 @@
 Support classes for the python server.
 '''
 from .Server_Thread import Server_Thread
+from .Misc import Client_Garbage_Collected
 from .Pipe import Pipe_Server, Pipe_Client